add: fixed #pr-33 feedbacks
