+++ src/legacy/containers/Root.tsx
@@ -4,15 +4,13 @@ import { MemoryRouter } from 'react-router-dom'
 import { ErrorModal as ReduxErrorModal } from '../../features/errorModal'
 import Routes from '../Routes'
 
-const Root = (): JSX.Element => {
-  return (
-    <>
-      <MemoryRouter>
-        <Routes />
-      </MemoryRouter>
-      <ReduxErrorModal />
-    </>
-  )
-}
+const Root = (): JSX.Element => (
+  <>
+    <MemoryRouter>
+      <Routes />
+    </MemoryRouter>
+    <ReduxErrorModal />
+  </>
+)
 
 export default Root
